requirements.txt
asgiref==3.5.2
Django==3.2
django-admin-rangefilter==0.8.5
django-modeltranslation==0.18.4
pdfkit==1.0.0
pytz==2022.1
sqlparse==0.4.2
